requirements.txt
pyaes==1.6.1
pyasn1==0.5.0
python-dotenv==1.0.0
rsa==4.9
Telethon==1.28.5
tqdm==4.65.0
